Profile pic and vault utility.
Adding the ability for a user to upload a user profile. Each user will have to have their own blob container due to the inefficiencies of Azure blob retrieval tools.
Built a utility tool to provide CRUD-like functionality for Azure Key Vaults (will expand to include the keys in the future).